fix(shiki): add theme-free core entry
Written by an AI agent on behalf of @atinux; not yet human-reviewed.
